+++ src/cookie.js
@@ -1,33 +1,33 @@
-import JSCookie from "js-cookie"
+import JSCookie from "js-cookie";
 
-const USER_ID = "__HORUS_USER_ID"
+const USER_ID = "__HORUS_USER_ID";
 
 function geUserID() {
-    let uid = JSCookie.get(USER_ID)
+    let uid = JSCookie.get(USER_ID);
 
     if (!uid) {
-        uid = "anonym." + (new Date().getTime() / 1000 + Math.random())
-        let host = location.host.split(".")
-        let domain = ["*"]
-        domain[2] = host.pop()
-        domain[1] = host.pop()
+        uid = "anonym." + (new Date().getTime() / 1000 + Math.random());
+        let host = location.host.split(".");
+        let domain = ["*"];
+        domain[2] = host.pop();
+        domain[1] = host.pop();
 
         JSCookie.set(USER_ID, uid, {
-            expires: 365 * 7
-        })
+            expires: 365 * 7,
+        });
         JSCookie.set(USER_ID, uid, {
             expires: 365 * 7,
-            domain: domain.join(".")
-        })
+            domain: domain.join("."),
+        });
     }
-    return uid
+    return uid;
 }
 
 function getAccountID() {
-    return JSCookie.get("account_id")
+    return JSCookie.get("account_id");
 }
 
 export default {
     geUserID,
-    getAccountID
-}
+    getAccountID,
+};