+++ src/listeners/hover.js
@@ -5,7 +5,8 @@ function Hover(event) {
         target = event.target,
         attr;
     while (target && target != document) {
-        attr = target.getAttribute("ho:hover") || target.getAttribute("ho-hover");
+        attr =
+            target.getAttribute("ho:hover") || target.getAttribute("ho-hover");
         if (attr) {
             go = true;
             break;
@@ -14,14 +15,15 @@ function Hover(event) {
     }
     if (!go) return;
 
-    let a = attr.split(':'),
-    eventName = a[0], desc = a[1] || (target.innerText || '').substr(0, 20)
+    let a = attr.split(":"),
+        eventName = a[0],
+        desc = a[1] || (target.innerText || "").substr(0, 20);
 
     return {
-        eventType: 'hover',
+        eventType: "hover",
         eventName: eventName,
         desc: desc,
-        xpath: getXPath(target)
+        xpath: getXPath(target),
     };
 }