+++ src/listeners/index.js
@@ -1,9 +1,9 @@
-import Click from './click.js'
+import Click from "./click.js";
 // import Link from './link.js'
-import Hover from './hover.js'
+import Hover from "./hover.js";
 // import History from './history.js'
 
 export default {
     Click,
     Hover,
-}
+};