+++ src/serialize.js
@@ -1,10 +1,8 @@
-import {
-    Base64
-} from "js-base64"
+import { Base64 } from "js-base64";
 
-function Serialize(schema){
-    if(typeof schema == "object") schema = JSON.stringify(schema)
-    return Base64.encodeURI(schema)
+function Serialize(schema) {
+    if (typeof schema == "object") schema = JSON.stringify(schema);
+    return Base64.encodeURI(schema);
 }
 
-export default Serialize
+export default Serialize;