+++ src/throttle.js
@@ -1,23 +1,23 @@
 function Debounce(fn, delay = 1000) {
-    let timer
+    let timer;
 
     return function (event) {
-        clearTimeout(timer)
+        clearTimeout(timer);
         timer = setTimeout(function () {
-            fn(event)
-        }, delay)
-    }
+            fn(event);
+        }, delay);
+    };
 }
 
 function Throttle(fn, delay = 1000) {
-    let last = 0
+    let last = 0;
     return function (event) {
-        let now = new Date().getTime()
+        let now = new Date().getTime();
         if (now - last > delay) {
-            fn(event)
-            last = now
+            fn(event);
+            last = now;
         }
-    }
+    };
 }
 
-export default Throttle
+export { Throttle as default, Debounce };