+++ test/horus.js
@@ -1,19 +1,18 @@
-import Assert from 'assert'
-import Horus from '../src/index.js'
+import Assert from "assert";
+import Horus from "../src/index.js";
 
-describe('Horus', function () {
+describe("Horus", function () {
+    it("init", function () {
+        let horus = new Horus();
 
-    it('init', function () {
-        let horus = new Horus()
+        Assert.equal(typeof horus, "object");
+        Assert.equal(horus.opt.url, "");
+        Assert.equal(horus.opt.project, "Empty");
+        Assert.equal(horus.opt.debug, false);
 
-        Assert.equal(typeof horus, 'object')
-        Assert.equal(horus.opt.url, '')
-        Assert.equal(horus.opt.project, 'Empty')
-        Assert.equal(horus.opt.debug, false)
-
-        horus.setConfig('url', 'url value')
-        Assert.equal(horus.opt.url, 'url value')
-        horus.setConfig({url: 'new url value'})
-        Assert.equal(horus.opt.url, 'new url value')
-    })
-})
+        horus.setConfig("url", "url value");
+        Assert.equal(horus.opt.url, "url value");
+        horus.setConfig({ url: "new url value" });
+        Assert.equal(horus.opt.url, "new url value");
+    });
+});